Add more adjustments
Trying out a bund of experiments that did not really help anything. Did record my notes about my various experiments to try and optimize the memory loads. Seems like the big issue is a TLB miss, during get, when the bucket contains the key.

My notes are in the testrun.md file.