Make caymap generics-aware
Caymap is now with a generic value type and it seems that it had no impact on the benchmarks.

Further I extended the benchmarks with differing sizes and it seems that caymap is on par with the builtin. For the small ranges the builtin is faster and for the midrange caymap is faster. I need to figure out why caymap is slower in the bigger end as well.